+++ packages/nextjs/config/chains/index.ts
@@ -1,0 +1,1 @@
+export * from "./xeneaTestnet";
