+++ packages/nextjs/config/scaffold.config.ts
@@ -1,3 +1,4 @@
+import { xeneaTestnet } from "./chains";
 import * as chains from "viem/chains";
 
 export type ScaffoldConfig = {
@@ -6,28 +7,6 @@ export type ScaffoldConfig = {
   alchemyApiKey: string;
   walletConnectProjectId: string;
   onlyLocalBurnerWallet: boolean;
-};
-
-/**
- * Xenea testnet config
- */
-export const xeneaTestnet: chains.Chain = {
-  id: 5555,
-  name: "Xenea Testnet",
-  nativeCurrency: {
-    decimals: 18,
-    name: "XCR",
-    symbol: "XCR",
-  },
-  rpcUrls: {
-    default: { http: ["https://rpc-kura.cross.technology"] },
-  },
-  blockExplorers: {
-    default: {
-      name: "XeneaScan",
-      url: "https://testnet.crossvaluescan.com/",
-    },
-  },
 };
 
 chains.arbitrum;
